+++ aztec-abyss-dimension/src/main/java/com/jrpetty/aztecabyss/engine/Ruleset.java
@@ -1,7 +1,9 @@
 package com.jrpetty.aztecabyss.engine;
 
 import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import net.minecraft.world.entity.EntityType;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;